requirements.txt
flet
python-dotenv
websocket
sqlmodel
flet-fastapi
pyjwt